backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
Pillow==8.3.1
requests==2.26.0
python-dotenv==0.19.0 
pony==0.7.16
pymysql==1.0.2